moved animal combinations to shared structure
