+++ RUFAS/routines/manure/pen_manure/manure_manager_pen.py
@@ -1,11 +1,11 @@
-from typing import List, NamedTuple
+from typing import List, NamedTuple, Any
 from typing import Set
 
 from RUFAS.routines.animal.life_cycle.animal_base import AnimalBase
 from RUFAS.routines.animal.life_cycle.cow import Cow
 from RUFAS.routines.animal.pen import Pen
 from RUFAS.routines.manure.pen_manure.pen_manure import PenManure
-from RUFAS.routines.animal.animal_combinations import AnimalCombination
+from ....shared_structures.animal_combinations import AnimalCombination
 
 
 class ManureManagerPen:
@@ -62,10 +62,10 @@ class ManureManagerPen:
 
         """
         self.id: int = pen.id
-        self.animals_in_pen = pen.animals_in_pen
-        self.num_animals = len(pen.animals_in_pen)
+        self.animals_in_pen: dict[int, Any] = pen.animals_in_pen
+        self.num_animals: int = len(pen.animals_in_pen)
         self.classes_in_pen: Set[str] = pen.classes_in_pen
-        self.animal_combination: Pen.AnimalCombination = pen.animal_combination
+        self.animal_combination: AnimalCombination = pen.animal_combination
 
         self.housing_type: str = pen.housing_type
         self.pen_type: str = pen.pen_type
@@ -76,9 +76,9 @@ class ManureManagerPen:
         self.manure_separator_after_digestion: str = pen.manure_separator_after_digestion
         self.manure_treatment: str = pen.manure_storage
 
-        self.manure = PenManure.get_instance(pen.manure, self.num_animals)
-        self.num_lactating_cows = self.count_lactating_cows(pen.animal_combination, pen.animals_in_pen)
-        self.num_stalls = pen.num_stalls
+        self.manure: PenManure = PenManure.get_instance(pen.manure, self.num_animals)
+        self.num_lactating_cows: int = self.count_lactating_cows(pen.animal_combination, pen.animals_in_pen)
+        self.num_stalls: int = pen.num_stalls
 
     @classmethod
     def count_lactating_cows(cls, animal_combination: AnimalCombination, animals_in_pen: List[AnimalBase]) -> int:
