+++ tests/animal_module_tests/test_pen.py
@@ -15,7 +15,7 @@ from RUFAS.routines.animal.life_cycle.heiferII import HeiferII
 from RUFAS.routines.animal.life_cycle.heiferIII import HeiferIII
 from RUFAS.routines.animal.manure.general_manure import AnimalManureExcretions
 from RUFAS.routines.animal.pen import Pen
-from RUFAS.routines.animal.animal_combinations import AnimalCombination
+from RUFAS.shared_structures.animal_combinations import AnimalCombination
 
 
 @pytest.fixture
