+++ tests/test_simulation_engine.py
@@ -85,10 +85,10 @@ def test_daily_simulation(mocker: MockerFixture) -> None:
     simulation_engine.field_manager = mocker.MagicMock()
     simulation_engine.weather = mocker.MagicMock()
     simulation_engine.time = mocker.MagicMock()
+    simulation_engine.time.simulation_day = 10
     simulation_engine.animal_manager.all_pens = mocker.MagicMock()
     simulation_engine.animal_manager.simulation_day = mocker.MagicMock()
 
-    patch_for_simulate_daily_manure_manager = mocker.patch("RUFAS.simulation_engine.simulate_daily_manure_manager")
     patch_for_daily_feed_routine = mocker.patch("RUFAS.simulation_engine.routines.daily_feed_routine")
     patch_for_advance_time = mocker.patch.object(simulation_engine, "_advance_time")
 
@@ -101,11 +101,7 @@ def test_daily_simulation(mocker: MockerFixture) -> None:
         simulation_engine.weather,
         simulation_engine.time,
     )
-    patch_for_simulate_daily_manure_manager.assert_called_once_with(
-        simulation_engine.manure_manager,
-        simulation_engine.animal_manager.all_pens,
-        simulation_engine.animal_manager.simulation_day,
-    )
+    simulation_engine.manure_manager.daily_update.assert_called_once_with(simulation_engine.animal_manager.all_pens, simulation_engine.time.simulation_day)
     simulation_engine.field_manager.daily_update_routine.assert_called_once_with(
         simulation_engine.weather, simulation_engine.time
     )
